+++ src/AdventureException.php
@@ -1,0 +1,4 @@
+<?php
+namespace Nereare\Shadows;
+
+class AdventureException extends \Exception {}