+++ src/InvalidAidException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Nereare\Shadows;
+
+use Nereare\Shadows\AdventureException;
+
+class InvalidAidException extends AdventureException {}