+++ src/NoAidException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Nereare\Shadows;
+
+use Nereare\Shadows\AdventureException;
+
+class NoAidException extends AdventureException {}