+++ src/NoUuidException.php
@@ -1,0 +1,4 @@
+<?php
+namespace Nereare\Shadows;
+
+class NoUuidException extends \Exception {}