+++ CHANGELOG.md
@@ -18,4 +18,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Pull request.
 - ReadMe file.
 - Changelog file.
-- ...
+- User Profile classes.
+- Adventure classes.